abstraction of table udf_data
git-svn-id: http://svn.ilias.de/svn/ilias/trunk@19138 21b2c9ec-7c21-0410-8b45-9bfb8ed2bfc5

+++ Services/User/classes/class.ilObjUserGUI.php
@@ -923,7 +923,7 @@ class ilObjUserGUI extends ilObjectGUI
 			{
 				if (substr($k, 0, 4) == "udf_")
 				{
-					$udf[(int) substr($k, 4)] = $v;
+					$udf[substr($k, 4)] = $v;
 				}
 			}
 			$userObj->setUserDefinedData($udf);
@@ -1118,7 +1118,7 @@ class ilObjUserGUI extends ilObjectGUI
 			{
 				if (substr($k, 0, 4) == "udf_")
 				{
-					$udf[(int) substr($k, 4)] = $v;
+					$udf[substr($k, 4)] = $v;
 				}
 			}
 			$this->object->setUserDefinedData($udf);
